backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
libsql-experimental>=0.0.50
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
